Refactor main into modular structure
