+++ README.md
@@ -8,3 +8,10 @@ Mitarbeiter Abwesenheiten visuallisieren und filtern
   Der Benutzer startet die App.
 
 
+
+## Struktur
+- `constants.py` hält zentrale Konstanten.
+- `data_utils.py` enthält Funktionen zum Laden und Aufbereiten der Daten.
+- `figures.py` generiert alle Diagramme.
+- `app.py` definiert Layout und Callbacks der Dash-App.
+- `main.py` startet lediglich den Server.
